+++ app/src/main/java/com/example/vinilos/MainActivity.kt
@@ -37,7 +37,7 @@ class MainActivity : AppCompatActivity() {
         // menu should be considered as top level destinations.
         appBarConfiguration = AppBarConfiguration(
             setOf(
-                R.id.nav_home, R.id.nav_gallery, R.id.nav_slideshow
+                R.id.nav_home, R.id.nav_albums, R.id.nav_collectors
             ), drawerLayout
         )
         setupActionBarWithNavController(navController, appBarConfiguration)
