+++ app/src/main/java/com/example/vinilos/ui/albums/AlbumsFragment.kt
@@ -1,4 +1,4 @@
-package com.example.vinilos.ui.slideshow
+package com.example.vinilos.ui.albums
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -7,11 +7,11 @@ import android.view.ViewGroup
 import android.widget.TextView
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.ViewModelProvider
-import com.example.vinilos.databinding.FragmentSlideshowBinding
+import com.example.vinilos.databinding.FragmentAlbumsBinding
 
-class SlideshowFragment : Fragment() {
+class AlbumsFragment : Fragment() {
 
-    private var _binding: FragmentSlideshowBinding? = null
+    private var _binding: FragmentAlbumsBinding? = null
 
     // This property is only valid between onCreateView and
     // onDestroyView.
@@ -22,14 +22,14 @@ class SlideshowFragment : Fragment() {
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View {
-        val slideshowViewModel =
-            ViewModelProvider(this).get(SlideshowViewModel::class.java)
+        val albumsViewModel =
+            ViewModelProvider(this).get(AlbumsViewModel::class.java)
 
-        _binding = FragmentSlideshowBinding.inflate(inflater, container, false)
+        _binding = FragmentAlbumsBinding.inflate(inflater, container, false)
         val root: View = binding.root
 
-        val textView: TextView = binding.textSlideshow
-        slideshowViewModel.text.observe(viewLifecycleOwner) {
+        val textView: TextView = binding.textAlbums
+        albumsViewModel.text.observe(viewLifecycleOwner) {
             textView.text = it
         }
         return root
