+++ app/src/main/java/com/example/vinilos/ui/albums/AlbumsViewModel.kt
@@ -1,10 +1,10 @@
-package com.example.vinilos.ui.gallery
+package com.example.vinilos.ui.albums
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
-class GalleryViewModel : ViewModel() {
+class AlbumsViewModel : ViewModel() {
 
     private val _text = MutableLiveData<String>().apply {
         value = "Fragmento Albumes"
