+++ app/src/main/java/com/example/vinilos/ui/collectors/CollectorsFragment.kt
@@ -1,4 +1,4 @@
-package com.example.vinilos.ui.gallery
+package com.example.vinilos.ui.collectors
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -7,11 +7,11 @@ import android.view.ViewGroup
 import android.widget.TextView
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.ViewModelProvider
-import com.example.vinilos.databinding.FragmentGalleryBinding
+import com.example.vinilos.databinding.FragmentCollectorsBinding
 
-class GalleryFragment : Fragment() {
+class CollectorsFragment : Fragment() {
 
-    private var _binding: FragmentGalleryBinding? = null
+    private var _binding: FragmentCollectorsBinding? = null
 
     // This property is only valid between onCreateView and
     // onDestroyView.
@@ -22,14 +22,14 @@ class GalleryFragment : Fragment() {
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View {
-        val galleryViewModel =
-            ViewModelProvider(this).get(GalleryViewModel::class.java)
+        val collectorsViewModel =
+            ViewModelProvider(this).get(CollectorsViewModel::class.java)
 
-        _binding = FragmentGalleryBinding.inflate(inflater, container, false)
+        _binding = FragmentCollectorsBinding.inflate(inflater, container, false)
         val root: View = binding.root
 
-        val textView: TextView = binding.textGallery
-        galleryViewModel.text.observe(viewLifecycleOwner) {
+        val textView: TextView = binding.textCollectors
+        collectorsViewModel.text.observe(viewLifecycleOwner) {
             textView.text = it
         }
         return root
