+++ app/src/main/java/com/example/vinilos/ui/collectors/CollectorsViewModel.kt
@@ -1,10 +1,10 @@
-package com.example.vinilos.ui.slideshow
+package com.example.vinilos.ui.collectors
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
-class SlideshowViewModel : ViewModel() {
+class CollectorsViewModel : ViewModel() {
 
     private val _text = MutableLiveData<String>().apply {
         value = "Fragmento Coleccionistas"
