+++ app/src/main/res/layout/fragment_collectors.xml
@@ -4,10 +4,10 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.gallery.GalleryFragment">
+    tools:context=".ui.collectors.CollectorsFragment">
 
     <TextView
-        android:id="@+id/text_gallery"
+        android:id="@+id/text_collectors"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:layout_marginStart="8dp"
