+++ app/src/main/res/menu/activity_main_drawer.xml
@@ -9,11 +9,11 @@
             android:icon="@drawable/baseline_stars_24"
             android:title="@string/menu_home" />
         <item
-            android:id="@+id/nav_gallery"
-            android:icon="@drawable/ic_menu_slideshow"
+            android:id="@+id/nav_albums"
+            android:icon="@drawable/ic_menu_collectors"
             android:title="@string/menu_albums" />
         <item
-            android:id="@+id/nav_slideshow"
+            android:id="@+id/nav_collectors"
             android:icon="@drawable/baseline_people_24"
             android:title="@string/menu_collectors" />
     </group>
