+++ app/src/main/res/navigation/mobile_navigation.xml
@@ -12,14 +12,14 @@
         tools:layout="@layout/fragment_home" />
 
     <fragment
-        android:id="@+id/nav_gallery"
-        android:name="com.example.vinilos.ui.gallery.GalleryFragment"
+        android:id="@+id/nav_albums"
+        android:name="com.example.vinilos.ui.albums.AlbumsFragment"
         android:label="@string/menu_albums"
-        tools:layout="@layout/fragment_gallery" />
+        tools:layout="@layout/fragment_albums" />
 
     <fragment
-        android:id="@+id/nav_slideshow"
-        android:name="com.example.vinilos.ui.slideshow.SlideshowFragment"
+        android:id="@+id/nav_collectors"
+        android:name="com.example.vinilos.ui.collectors.CollectorsFragment"
         android:label="@string/menu_collectors"
-        tools:layout="@layout/fragment_slideshow" />
+        tools:layout="@layout/fragment_collectors" />
 </navigation>
